pyproject.toml
[project]
name = "IDF"
version = "0.0.0"
dependencies = [
    "python-dotenv==1.2.2",
    "requests==2.32.5",
    "clickhouse-connect==0.13.0"
]
